+++ components/Common/Navigation/index.js
@@ -1,16 +1,19 @@
-import { StyledLink } from "../../styles";
+import { StyledLink } from "../../../styles";
 import useLocalStorageState from "use-local-storage-state";
 import styled from "styled-components";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faListCheck } from "@fortawesome/free-solid-svg-icons";
+import { faBoxArchive } from "@fortawesome/free-solid-svg-icons";
 
 export default function Navigation() {
   const [selectedNavItem, setSelectedNavItem] = useLocalStorageState(
-    "selectedNavItem",
+    "projects",
     { defaultValue: "" }
   );
 
   function handleNavOnClick(event) {
     event.preventDefault();
-    setSelectedNavItem(event.target.innerText.toLowerCase());
+    setSelectedNavItem(event.target.innerText?.toLowerCase());
   }
 
   return (
@@ -20,6 +23,7 @@ export default function Navigation() {
         className="projects"
         selectedNavItem={selectedNavItem}
       >
+        <FontAwesomeIcon icon={faListCheck}>Projects</FontAwesomeIcon>
         <StyledLink href="/">Projects</StyledLink>
       </NavItem>
 
@@ -28,6 +32,7 @@ export default function Navigation() {
         className="stocks"
         selectedNavItem={selectedNavItem}
       >
+        <FontAwesomeIcon icon={faBoxArchive}> Stocks</FontAwesomeIcon>
         <StyledLink href="/stock">Stocks</StyledLink>
       </NavItem>
     </Nav>
@@ -40,15 +45,18 @@ const Nav = styled.nav`
   display: flex;
   justify-content: space-evenly;
   height: 4rem;
-  font-size: 1.5rem;
+  font-size: 1rem;
   font-weight: 400;
-  line-height: 4rem;
   width: 100%;
   text-align: center;
   background-color: #ffffff;
   text-decoration: none;
 `;
 const NavItem = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 0.6rem;
   width: 50%;
   background-color: #fff;
   color: ${({ selectedNavItem, className }) =>
