+++ components/Project/handleProjectRestructure/index.js
