+++ pages/_app.js
@@ -4,7 +4,7 @@ import { useRouter } from "next/router";
 import { useState } from "react";
 import { SWRConfig } from "swr";
 import useSWR from "swr";
-import { handleProjectRestructure } from "../components/handleProjectRestructure";
+import { handleProjectRestructure } from "../components/Project/handleProjectRestructure";
 import { Lato } from "@next/font/google";
 import { uid } from "uid";
