+++ pages/stock/index.js
@@ -1,10 +1,10 @@
-import YarnStockList from "../../components/YarnStockList";
-import Navigation from "../../components/Navigation";
-import Heading from "../../components/Heading";
-import AddButton from "../../components/AddButton";
+import YarnStockList from "../../components/Stock/YarnStockList";
+import Navigation from "../../components/Common/Navigation";
+import Heading from "../../components/Common/Heading";
+import AddButton from "../../components/Common/AddButton";
 import { Main } from "../../styles";
 import { useState } from "react";
-import YarnStockForm from "../../components/YarnStockForm";
+import YarnStockForm from "../../components/Stock/YarnStockForm";
 
 export default function YarnsStockPage() {
   const [addYarnStockStatus, setAddYarnStockStatus] = useState(false);
